requirements.txt
flask>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
flask-sqlalchemy>=3.0.0
werkzeug>=2.3.0

